move declaration of DarkBremInteraction into its own header
